refactor: move munging of args into DynamicUdfInvoker
